+++ src/pages/main/components/WithAnswer.tsx
@@ -1,6 +1,10 @@
 import BUNDLE_WITH_ANSWER_IMAGE from "@/shared/assets/pouch/pouch_with_answer.png";
 import Bundle from "./Bundle";
 import { MainPageInfo } from "@/types/main-page";
+import { Button } from "@/components/ui/button";
+import { ReflectionButton } from "./ReflectionButton";
+import ShareButton from "@/components/share/ShareButton";
+import { useHistory } from "react-router-dom";
 
 type Props = {
   userId: number;
@@ -8,12 +12,18 @@ type Props = {
 };
 
 export default function WithAnswer({
+  userId,
   mainPageInfo: {
     totalCount: answerCount,
     answerContent: previewMessage,
     content: questionContent,
   },
 }: Props) {
+  const history = useHistory();
+  const handleClick = () => {
+    history.push("/answer-result");
+  };
+
   return (
     <div className="flex flex-col items-center">
       <span className="text-white text-h2 mb-6">
@@ -33,6 +43,19 @@ export default function WithAnswer({
       <span className="font-nanum-dahaengce mb-4 text-white text-xl">
         누구의 쪽지일까요? 지금 열어보세요!
       </span>
+      <footer className="w-full">
+        <Button
+          className="mb-2 w-full"
+          children="열어보기"
+          onClick={handleClick}
+        />
+        <div className="flex w-full">
+          <ReflectionButton userId={userId} />
+          <ShareButton variant={"outline"} className="w-full ml-2">
+            공유
+          </ShareButton>
+        </div>
+      </footer>
     </div>
   );
 }
